requirements.txt
asgiref==3.6.0
Django==4.1.5
django-filter==22.1
django-model-utils==4.3.1
djangorestframework==3.14.0
Markdown==3.4.1
psycopg2-binary==2.9.5
pytz==2022.7.1
sqlparse==0.4.3
types-pkg-resources==0.1.3
Unipath==1.1